+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Support both datatable and docstring ([#136](https://github.com/cucumber/javascript-core/pull/136))
 
 ## [0.9.0] - 2026-06-11
 ### Changed